requirements.txt

langchain-community
langchain-core
langgraph
streamlit
pydantic
openai
pandas
python-dotenv
